backend/requirements.txt
# LangGraph 核心依赖
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# HTTP 客户端
httpx>=0.24.0
requests>=2.31.0

# API 服务器
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# 环境变量
python-dotenv>=1.0.0

# OpenAI 客户端
openai

# 数据库 - PostgreSQL
psycopg2-binary>=2.9.0

# 向量数据库 - Milvus
pymilvus>=2.5.0

# Token counting
tiktoken>=0.5.0

# 加密（namespace 密码加密保留）
cryptography>=45.0.7,<47

# 文件上传
python-multipart

# 阿里云服务
dashscope
oss2
alibabacloud_oss_v2

# 文档解析
pageindex
pymupdf
python-docx
